--- back-end/src/main/java/com/spring/mydiv/Dto/Travel.java
@@ -1,16 +1,0 @@
-package com.spring.mydiv.Dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@NoArgsConstructor
-@RequiredArgsConstructor
-public class Travel {
-	private int Travel_id;
-	@Setter @NonNull
-	private String TravelName;
-}
