--- back-end/src/main/java/com/spring/mydiv/Mapper/TravelMapper.java
@@ -1,13 +1,0 @@
-package com.spring.mydiv.Mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-
-import com.spring.mydiv.Dto.Travel;
-
-
-@Mapper
-public interface TravelMapper {
-	
-	public int insertTravelName(Travel travel);
-
-}
