--- back-end/src/main/java/com/spring/mydiv/Service/PersonService.java
@@ -1,13 +1,0 @@
-package com.spring.mydiv.Service;
-
-import java.util.List;
-
-import com.spring.mydiv.Dto.Person;
-
-public interface PersonService {
-	
-	List<Person> CreatePerson(Person person);
-	
-	Person ShowPerson(Person person);
-
-}
