--- back-end/src/main/java/com/spring/mydiv/Service/TravelService.java
@@ -1,9 +1,0 @@
-package com.spring.mydiv.Service;
-
-import com.spring.mydiv.Dto.Travel;
-
-public interface TravelService {
-	//fin_0330
-	String CreateTravel(Travel travel);
-
-}
