--- back-end/src/main/java/com/spring/mydiv/Try03301530Application.java
@@ -1,14 +1,0 @@
-package com.spring.mydiv;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Try03301530Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Try03301530Application.class, args);
-		System.out.print("fin");
-	}
-
-}
