--- back-end/src/main/resources/sqlmap/Travel_SQL.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-        
-<mapper namespace="com.spring.mydiv.Mapper.TravelMapper">
-
-	<insert id="insertTravelName" parameterType="travel" useGeneratedKeys="true" keyProperty="Travel_id">  
-	    <![CDATA[
-	    insert into travelDB (TravelName)
-	    values (#{TravelName})
-	    ]]>
-	</insert>
-
-</mapper>
