--- back-end/src/test/java/com/spring/mydiv/Try03301530ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.spring.mydiv;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Try03301530ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
