+++ src/main/java/com/spring/mydiv/Dto/Event.java
@@ -5,14 +5,16 @@ import java.util.ArrayList;
 import java.util.Date;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 @Getter
-@RequiredArgsConstructor
+@Setter
+@NoArgsConstructor
 public class Event {
-	
+
 	@NonNull
 	private String EventName; //in eventDB
 	@NonNull
@@ -20,23 +22,32 @@ public class Event {
 	@NonNull
 	private Person Payer; //in eventDB
 	@NonNull
-	private Date EventDate; //in eventDB
+	private String TravelName;
 	@NonNull
 	private int Price; //in eventDB
 	@NonNull
-	private String TravelName;
+	private Date EventDate; //in eventDB
 	
+	private int Event_id;
 	private String Participant2String; //in eventDB
-	@Setter
+	private String PayerName;
 	private double DividePrice;
-	@Setter
 	private double GetPrice;
 	
-	public void SetInfo() {
-		this.setDividePrice((double)this.Price / this.Participant.size());
-		this.setGetPrice(this.DividePrice * (this.Participant.size()-1));
+	public Event(String EventName, ArrayList<Person> Participant,
+			Person Payer, String TravelName, int Price, Date EventDate) {
+		this.setEventName(EventName);
+		this.setParticipant(Participant);
+		this.setPayer(Payer);
+		this.setTravelName(TravelName);
+		this.setPrice(Price);
+		this.setEventDate(EventDate);
+		
+		this.setPayerName(Payer.getPersonName());
+		this.setDividePrice((double)Price / Participant.size());
+		this.setGetPrice(this.DividePrice * (Participant.size()-1));
 		for(Person parti : Participant) {
-			Participant2String = String.join(Participant2String, " ", parti.getPersonName());
+			this.Participant2String += parti.getPersonName() + " " ;
 		}
 	}
