+++ src/main/java/com/spring/mydiv/Dto/Person.java
@@ -3,14 +3,16 @@ package com.spring.mydiv.Dto;
 import java.util.ArrayList;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
-@Getter @Setter
+@Getter
+@Setter
+@NoArgsConstructor
 @RequiredArgsConstructor
 public class Person {
-	
 	@NonNull
 	private String PersonName; //in personDB
 	@NonNull
@@ -18,6 +20,7 @@ public class Person {
 	@NonNull
 	private String TravelName; //in personDB
 	
+	private int Person_id;
 	private double sumSend;
 	private double sumGet;
 	private double Difference;	
@@ -25,5 +28,4 @@ public class Person {
 	private ArrayList<Event> willSend = new ArrayList<>();
 	private ArrayList<Event> willGet = new ArrayList<>();
 	private String Role;
-	
 }
