+++ src/main/java/com/spring/mydiv/Dto/Travel.java
@@ -1,11 +1,13 @@
 package com.spring.mydiv.Dto;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 @Getter
+@NoArgsConstructor
 @RequiredArgsConstructor
 public class Travel {
 	private int Travel_id;
