+++ src/main/java/com/spring/mydiv/Service/EventService.java
@@ -3,13 +3,11 @@ package com.spring.mydiv.Service;
 import java.util.List;
 
 import com.spring.mydiv.Dto.Event;
+import com.spring.mydiv.Dto.Person;
 
 public interface EventService {
-
-	long getTravelPeriod(Event event);
-	long getEventNum(Event event);
 	List<Event> CreateEvent(Event event);
-	Event ShowEvent(Event event);
-	List<String> ShowWhoInTravel(Event event);
-
+	boolean get_difference(Event event, Person person);
+	void SetRole(String travelName);
+	boolean UpdateRole(Person person);
 }
