+++ src/main/java/com/spring/mydiv/Service/EventServiceImpl.java
@@ -1,63 +1,96 @@
 package com.spring.mydiv.Service;
-
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
-import com.spring.mydiv.Dao.EventDao;
-import com.spring.mydiv.Dao.PersonDao;
 import com.spring.mydiv.Dto.Event;
 import com.spring.mydiv.Dto.Person;
-import com.spring.mydiv.Dto.Travel;
-
-public class EventServiceImpl implements EventService {
+import com.spring.mydiv.Mapper.EventMapper;
+import com.spring.mydiv.Mapper.PersonMapper;
+@Service
+public class EventServiceImpl implements EventService{
 	@Autowired
-	private EventDao eventDao;
+	EventMapper eventMapper;
 	@Autowired
-	private PersonDao personDao;
+	PersonMapper personMapper;
 	
-	@Autowired
-	private PersonService personService;
-	@Autowired
-	private TravelService travelService;
-	
-	//----------create----------//
-	@Override
 	public List<Event> CreateEvent(Event event) {
-		int affectRowCount = this.eventDao.insert(event);
-	    if (affectRowCount ==  1) { // 1 = success
+		// event.SetInfo();
+		//System.out.println("setinfo success");
+		int affectRowCount = this.eventMapper.insertEventInfo(event);
+		//System.out.println("insert event success");
+		if (affectRowCount ==  1) { // 1 = success
 	    	for(Person parti : event.getParticipant()) {
-	    		personService.get_difference(event, parti);
+	    		get_difference(event, parti);
+	    		//System.out.println("get diff success");
 	    	}
-	    	String travelName = event.getTravelName();
-	    	travelService.SetRole(travelName);
-	    	return this.eventDao.getEventBriefly(travelName);
+	    	String TravelName = event.getTravelName();
+	    	SetRole(TravelName);
+	    	//System.out.println("setrole success");
+	    	return this.eventMapper.getEventBriefly(TravelName);
 	    }
 	    return null;
 	}
 	
-	@Override
-	public long getTravelPeriod(Event event) {
-		String travelName = event.getTravelName();
-		return this.eventDao.getTravelPeriod(travelName);
+	public boolean get_difference(Event event, Person person) {
+		if (person.getPersonName() == event.getPayer().getPersonName()){
+			//System.out.println(person.getPersonName());
+			person.setSumGet(person.getSumGet()+event.getGetPrice());			
+		} else {
+			person.setSumSend(person.getSumSend()+event.getDividePrice());
+		}
+		person.setDifference(person.getSumGet() - person.getSumSend());
+		//System.out.println(person.getDifference());
+		//System.out.println(person.getPerson_id());
+		int affectRowCount = this.personMapper.updatePersonMoney(person);
+		return affectRowCount == 1;
 	}
 	
-	@Override
+	public void SetRole(String travelName) {
+		double max_difference = 0;
+		Person Leader = null;
+		
+		List<Person> personALL = this.personMapper.getWhoInTravelAll(travelName);
+		for(Person person : personALL) {
+			if(person.getDifference() < 0) {
+				person.setRole("SENDER");
+			}
+			else {
+				person.setRole("GETTER");
+				if (max_difference < person.getDifference()) {
+					Leader = person;
+					max_difference = person.getDifference();
+				}
+			}
+		}
+		Leader.setRole("LEADER");
+		
+		for(Person person : personALL) {
+			if (!UpdateRole(person)) {
+				System.out.println("Fail!");
+			}
+		}	
+	}
+
+	public boolean UpdateRole(Person person) {
+		int affectRowCount = this.personMapper.updatePersonRole(person);
+		return affectRowCount == 1;
+	}
+	
 	public long getEventNum(Event event) {
 		String travelName = event.getTravelName();
-		return this.eventDao.getEventNum(travelName);
+		return this.eventMapper.getEventNum(travelName);
 	}
 	
 	//----------detail----------//
-	@Override
 	public Event ShowEvent(Event event) {
-		return this.eventDao.getEventDetail(event);
+		return this.eventMapper.getEventDetail(event);
 	}
 	
 	//----------for create----------//
-	@Override
-	public List<String> ShowWhoInTravel(Event event){
-		return this.personDao.getWhoInTravelName(event);
+	public List<String> ShowWhoInTravel(String TravelName){
+		return this.personMapper.getWhoInTravelName(TravelName);
 	}
 	
 }
