--- src/main/resources/application.properties
@@ -1,11 +1,0 @@
-# DB
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/example?serverTimezone=UTC&characterEncoding=UTF-8
-spring.datasource.username=20205149
-spring.datasource.password=20205149
-
-# xml alias -> packageX
-mybatis.type-aliases-package=com.spring.mydiv.Dto
-
-# Mybatis mapper
-mybatis.mapper-locations= sqlmap/*.xml
